fix : user cash Integer->Long 변경

+++ core/src/main/java/io/eagle/entity/User.java
@@ -47,5 +47,5 @@ public class User {
     private String refreshToken;
 
     @Column
-    private Integer cash;
+    private Long cash;
 }
